services/recommendationservice/requirements.txt
#
# This file is for AWS-based recommendation service
# Updated from GCP to AWS dependencies
#
# Core AWS dependencies
boto3>=1.34.0
aws-xray-sdk>=2.12.0

# gRPC dependencies
grpcio>=1.76.0
grpcio-health-checking>=1.76.0

# Logging
python-json-logger>=4.0.0

# HTTP client
requests>=2.32.0

# OpenTelemetry tracing
opentelemetry-distro>=0.41b0
opentelemetry-instrumentation-grpc>=0.59b0
opentelemetry-exporter-otlp-proto-grpc>=1.38.0
